% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{beer}
\alias{beer}
\title{Beer classification data}
\usage{
data(beer)
}
\value{
A tibble
}
\description{
Alcohol percentage, colour rating for
stouts and ales.
}
\examples{
data(beer)
}
