% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{beer_full}
\alias{beer_full}
\title{Beer classification data}
\usage{
data(beer_full)
}
\value{
A tibble
}
\description{
Alcohol percentage, colour rating for
stouts, IPAs and ales.
}
\examples{
data(beer_full)
}
