% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{beer_test}
\alias{beer_test}
\title{Beer test data}
\usage{
data(beer_test)
}
\value{
A tibble
}
\description{
Alcohol percentage, colour rating for
5 stouts and ales.
}
\examples{
data(beer_test)
}
