% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_plot.R
\name{boundary_plot}
\alias{boundary_plot}
\title{Convience function for creating boundary plots}
\usage{
boundary_plot(
  model,
  x,
  y,
  z,
  xlab = all.vars(formula(model))[2],
  ylab = all.vars(formula(model))[3],
  jitter = TRUE,
  xlim = range(x),
  ylim = range(y)
)
}
\arguments{
\item{model}{A caret classification model}

\item{x}{Data}

\item{y}{Data}

\item{z}{Observed values}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}

\item{jitter}{Default \code{TRUE}. Should the points be jittered.}

\item{xlim}{x axis limit}

\item{ylim}{y axis limit}
}
\description{
Convience function for creating boundary plots
}
