% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{car}
\alias{car}
\title{Price data for different types of cars}
\source{
https://www.kaggle.com/hellbuoy/car-price-prediction
}
\usage{
data(car)
}
\value{
A tibble
}
\description{
Price data for different types of cars
}
\examples{
data(car)
}
