% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\docType{data}
\name{default}
\alias{default}
\title{The default data set}
\usage{
data(default)
}
\value{
A data frame
}
\description{
Contains loan data from the from the {ISLR} package. It contains
information on credit card debt for 10,000 people. Variables include
whether they've defaulted, their yearly income, current credit card
balance and whether they're a student.
}
\examples{
data(default)
}
