% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{gapminder}
\alias{gapminder}
\title{The full gapminder data set}
\source{
https://www.gapminder.org/
}
\usage{
data(gapminder)
}
\value{
A tibble
}
\description{
Measurements lifeExp, pop and gdpPercap for countries around the world
ranging from 1950 to 2020.
}
\examples{
data(gapminder)
}
