% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{india}
\alias{india}
\title{Gapminder data for india}
\source{
https://www.gapminder.org/
}
\usage{
data(india)
}
\value{
A tibble
}
\description{
Measurements lifeExp, pop and gdpPercap for India
ranging from 1950 to 2010.
}
\examples{
data(india)
}
