% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{india_full}
\alias{india_full}
\title{Full gapminder data for india}
\source{
https://www.gapminder.org/
}
\usage{
data(india_full)
}
\value{
A tibble
}
\description{
Measurements lifeExp, pop and gdpPercap for India
ranging from 1950 to 2019.
}
\examples{
data(india_full)
}
