% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{india_test}
\alias{india_test}
\title{Gapminder test data for india}
\source{
https://www.gapminder.org/
}
\usage{
data(india_test)
}
\value{
A tibble
}
\description{
Measurements lifeExp, pop and gdpPercap for India
ranging from 2011 to 2020.
}
\examples{
data(india_test)
}
