% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{lifeExp}
\alias{lifeExp}
\title{Life Expectancy data for countries around the world}
\usage{
data(lifeExp)
}
\value{
A tibble
}
\description{
Life Expectancy data for countries around the world
}
\examples{
data(lifeExp)
}
