% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_example.R
\docType{data}
\name{mixture_example}
\alias{mixture_example}
\title{The mixture example data set}
\description{
Mixture example data from the now archived ElemStatLearn package
}
\examples{
data(mixture_example)
}
