% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{penguins}
\alias{penguins}
\title{Trimmed penguins data}
\source{
The palmer penguins package
}
\usage{
data(penguins)
}
\value{
A tibble
}
\description{
Trimmed penguins data
}
\examples{
data(penguins)
}
