% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advertising.R
\docType{data}
\name{advertising}
\alias{advertising}
\title{The advertising data set}
\value{
A data.frame
}
\description{
200 measurements of advertising expenditure in each of
three domains, TV, Radio and Newspaper together with Sales
of the product being advertised.
}
\examples{
data(advertising)
}
